pyproject.toml
[project]
name = "fake-news-detector"
version = "1.0.0"
description = "Fake news detection using fine-tuned BERT model"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "newsapi-python>=0.2.7",
    "beautifulsoup4>=4.12.0",
    "serpapi>=0.1.5",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "httpx>=0.25.1",
]
